backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15,<4
django-cors-headers>=4.3,<5
